refactor(utils): split class, ordered map, string buffer
Move class, ordered map, and string buffer utilities into separate
modules for better organization and maintainability. Update all
references to use the new modules. This change improves code clarity
and makes future extensions easier.
