+++ lua/CopilotChat/ui/spinner.lua
@@ -1,6 +1,7 @@
 local notify = require('CopilotChat.notify')
 local utils = require('CopilotChat.utils')
-local class = utils.class
+local class = require('CopilotChat.utils.class')
+
 local spinner_frames = {
   '⠋',
   '⠙',
