+++ lua/CopilotChat/utils.lua
@@ -28,81 +28,6 @@ M.curl_args = {
   },
 }
 
----@class Class
----@field new fun(...):table
----@field init fun(self, ...)
-
---- Create class
----@param fn function The class constructor
----@param parent table? The parent class
----@return Class
-function M.class(fn, parent)
-  local out = {}
-  out.__index = out
-
-  local mt = {
-    __call = function(cls, ...)
-      return cls.new(...)
-    end,
-  }
-
-  if parent then
-    mt.__index = parent
-  end
-
-  setmetatable(out, mt)
-
-  function out.new(...)
-    local self = setmetatable({}, out)
-    fn(self, ...)
-    return self
-  end
-
-  function out.init(self, ...)
-    fn(self, ...)
-  end
-
-  return out
-end
-
----@class OrderedMap<K, V>
----@field set fun(self:OrderedMap, key:any, value:any)
----@field get fun(self:OrderedMap, key:any):any
----@field keys fun(self:OrderedMap):table
----@field values fun(self:OrderedMap):table
-
---- Create an ordered map
----@generic K, V
----@return OrderedMap<K, V>
-function M.ordered_map()
-  return {
-    _keys = {},
-    _data = {},
-    set = function(self, key, value)
-      if not self._data[key] then
-        table.insert(self._keys, key)
-      end
-      self._data[key] = value
-    end,
-
-    get = function(self, key)
-      return self._data[key]
-    end,
-
-    keys = function(self)
-      return self._keys
-    end,
-
-    values = function(self)
-      local result = {}
-      for _, key in ipairs(self._keys) do
-        table.insert(result, self._data[key])
-      end
-      return result
-    end,
-  }
-end
-
 --- Convert arguments to a table
 ---@param ... any The arguments
 ---@return table
@@ -119,39 +44,6 @@ function M.to_table(...)
     end
   end
   return result
-end
-
----@class StringBuffer
----@field add fun(self:StringBuffer, s:string)
----@field set fun(self:StringBuffer, s:string)
----@field tostring fun(self:StringBuffer):string
-
---- Create a string buffer for efficient string concatenation
----@return StringBuffer
-function M.string_buffer()
-  return {
-    _buf = { '' },
-
-    add = function(self, s)
-      table.insert(self._buf, s)
-      -- Keep track of lengths to know when to merge
-      for i = #self._buf - 1, 1, -1 do
-        if #self._buf[i] > #self._buf[i + 1] then
-          break
-        end
-        self._buf[i] = self._buf[i] .. table.remove(self._buf)
-      end
-    end,
-
-    set = function(self, s)
-      self._buf = { s }
-    end,
-
-    -- Get final string
-    tostring = function(self)
-      return table.concat(self._buf)
-    end,
-  }
 end
 
 --- Writes text to a temporary file and returns path
